Skip 'test_filereader_fails_to_open_file' permissions can be ignored
If the current user can ignore file permissions, the
`test_filereader_fails_to_open_file` test will not work correctly and
the error will not be the expected one. We skip the test in this case as
there is no reliable way to force an `OSError` in the C++ layer.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>