Add a mutex to the record reader to protect for concurrent access
Getting the stack trace of an allocation at the same time a thread is
pulling new allocation records is not a safe operation as the thread
pulling new allocations can mutate shared resources (the map of frames)
at the same time the thread that is getting allocations is reading from
these shared resources. To avoid this, we need a mutex to serialize
manipulation of said resources.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: László Kiss Kollár <lkisskollar@bloomberg.net>