Stop capturing a free port eagerly
The free port is only necessary when the user is trying to run in live
mode and has not provided their own port number.

A random port number was also presented in the documentation, since
`sphinx.ext.argparse` could only see an integer as the default for the
`--live-port` option. Since it is set up to show defaults when they are
non-None, every documentation build would pick a random free port and
use it.