Fix flaky subcommand tests
In `test_main.py`, when checking whether the pensieve command has been
executed correctly, we look at the contents of the generated HTML to
check for an allocation in the `json/tool.py` file was recorded.
Depending how the test was executed, we might not see an allocation
there, which can causes test failures.

Given that the these tests check command line parsing and output
generation, it is more reliable to check for the executed command line
in the output HTML.