Don't print error message if the reader reached EOF
We are incorrectly printing an error message for a failed close if the
`FileSource` reaches EOF and any of the read operations failed because
of it. This makes us incorrectly identify the error bits in the file
stream as coming from the `close()` call. To avoid this, clean the error
status in the file stream before attempting to close.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pradyun Gedam <pgedam@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>