Fix flaky test in `TestLiveSubcommand`
The `test_live_tracking_server_when_client_disconnects` test wants to
check that the tracker quits as soon as the client exits due to a 'q'
keypress. However, the server side will keep running until
the tracked process terminates, and this can cause issues if it does not
complete in the 10 seconds limit.

As the server prints the expected message immediately upon the client's
disconnection, we don't have to raise if the timeout is hit.