Add a command to start a pensieve live process in the background
Is quite cumbersome to have to prepare two different windows to start a
pensieve live process and attach to it. In most cases, users will like
to start the process in the background and attach inmediately to it,
ignoring any output that the process emmits.

For this reason, this commit does the following:

* Add a new mode that starts a process in the background and attaches to
  it.
* Names the new command `pesieve run  --live`.
* Renames the old `pensieve run --live` to `pensieve run --live-remote`.