Optimize the number of allocations we feed to Rich
Rich spends a considerable amount of time processing the records we pass
to display on the live view, even though depending on the terminal size,
most of the records won't be drawn on screen. We can optimize this by
checking the size of the terminal and using that as the maximum number
of records to pass to Rich.