Always return a value in `get_current_snapshot`
If the caller checks `is_active` before invoking `get_current_snapshot`,
to make sure that the background thread is still available, it can
happen that between the check, and requesting the snapshot, the
background thread gets destroyed. To avoid this, we can return an empty
list in `get_current_snapshot` if the background thread is unavailable.
This removes the need for the client to call `is_active`.