Shut down tracker process on exception in parent
When running in live mode, we have to ensure that the tracker child
process will be terminated if an exception is raised in the parent
process. An example where this causes issues is if the program is
running detached from a terminal and it receives a SIGINT: the child
will not receive the signal and it will keep running.