Ensure subprocesses are terminated in tests
We get resource warnings under the debug version of CPython about
subprocesses we haven't stopped in some of our tests. If we just kill
the process, we can still get the warning as the process might linger
for a while as a zombie until the parent reaps the child. For this
reason, add a wait with a timeout to ensure that the process is cleaned
up properly.

Authored-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: László Kiss Kollár <lkisskollar@bloomberg.net>