Add a summary reporter
Add a new `pensieve summary` subcommand and reporter that provides a
quick view on some statistics of the highest watermark allocation
records as well as a table showing different statistics and
relationships among functions that allocate memory.

This will be the same view as the live `TUI` provides but for static
files.