Ensure that the filtered nodes in the flamegraph are not mutated
Currently, we are mutating the "children" attribute of the flamegraph
nodes when we filter them for interesting frames. This has the side
effect that unmarking the checkbox for "interesting frames" won't show
back the full collection as the mutated version has eliminated a lot of
the original data.

To avoid this problem, make a copy of the nodes when attaching the new
children.