Log to stderr from C++ instead of to `logging`
This fixes a potential deadlock (we can't be certain that no thread is
holding the GIL and waiting for our thread to load a shared library, so
it's unsafe for us to try to acquire the GIL while loading a shared
library). It also incidentally fixes a deadlock when subinterpreters are
used, and it's faster and simpler.

Authored-by: Matt Wozniski <mwozniski@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>