Give frame PUSH and POP distinct record types
Previously we had one record type for frame actions, and a flag inside
that record to indicate whether the action was a PUSH or a POP.

Instead, make PUSH and POP into distinct record types. This prevents us
from needing to compute the frame ID for POP records (which never
needed or used it), and saves us the cost of storing the extra flag on
each frame action by encoding the 1 extra bit of information (whether
the action is a push or pop) into the record type.

In a quick benchmark, this made the output file from the `test_asyncio`
test in the Python test suite about 36% smaller with native tracking
disabled.