Detect when our TLS vector is created or destroyed
Wrap our thread local `vector` tracking the Python stack in a class that
sets another TLS pointer to the `vector` once it has been (lazily)
constructed in a new thread, and unsets it once the `vector` has been
destroyed.

When this new thread-local POD pointer is not set, it means that either
the thread-local `vector` tracking the Python stack has not been created
or has already been destroyed for the current thread. If it is set, it
means that the vector is still in scope and we can safely access it,
even from below `trackAllocation` and `trackDeallocation`.