Lazily emit frame PUSH and POP records
Rather writing a PUSH record as soon as a Python function is entered and
a POP record as soon as control returns from that function, only emit
them at the point where an allocation occurs below the Python function.
This avoids writing unnecessary PUSH and POP records in cases where no
allocation occurs below a called function, which happens very commonly
when using the `pymalloc` allocator.