Fix one of the problems detected by the debug macros
We are technically using a vector iterator pass the end. Although this
is undefined behavior, we are using only the vector's buffer so we don't
need the end iterator to be exact. To avoid running into undefined
behaviour resize the vector with 0s to make it think is always full.