tests: Remove an unreliable assertion
This test has been assuming that, after the client disconnects, the
server will discover that, print a message, and disable tracking. The
server will only discover that if it needs to flush its write buffer,
though, and if tracking ends before the write buffer is flushed, the
message will never be printed.

Rather than depend on any implementation details about when flushes
occur or whether a message is printed when there is data remaining in
the buffer when tracking ends, just remove this assertion.

The remaining assertion, that we never complain about the failing 'send'
call, is what we're more interested in anyway.