Stop emitting push and pop records on failure
Previously we were emitting these records in a loop, and if a write
failure occurred we wouldn't break out of the loop even though we had
deactivated tracking.

To prevent this message from being printed repeatedly unnecessarily,
check whether writing the output succeeded, and break out of the output
loops otherwise.