Ensure that the writer buffer is flushed when we dump memory maps
When we are dumping the memory maps to the writer, if there are any
records already in the writer's buffer it can happen that, as we are
using methods that do not write directly to the buffer, those records
that were sitting previously are written in the middle of the memory map
records.

To avoid this situation from happening, ensure that the buffer is
flushed when we acquire the writer lock explicitly.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: Matt Wozniski <mwozniski@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>