Replace json.tool with something performing deterministic allocations
When running some tests, the usage of `json.tool` as the script to
profile can lead to some non-deterministic allocation patterns. To
validate allocations in `test_main.py` we switch to use something which
performs a native allocation as these test are for testing the CLI logic
and they don't need to validate allocations.