Ensure that the `Destination` classes come from the same module
We have been seeing some errors due to the fact that when Cython makes a
relative import, the module of the classes is set as
`pensieve.<CLASSNAME>` but when the `__init__.py` file imports the
classes the module is set as `bloomberg.pensieve.<CLASSNAME>`. This
causes the `isinstance` calls in the extension to fail.

To avoid this problem, ensure that the classes imported via the top
level module are the same as the extension is using.