Add a badge in web reports with the Python allocator that was used
Interpreting results of some of the reports can be challenging when the
pymalloc allocator is used. This is because as the pymalloc allocator
handles its own heap, memory patterns will not correspond to the
creation patterns and destruction of objects. This can be specially bad
in the reports with `--leaks` as the reported leaks may be a bigger
subset due to memory getting allocated but not freed by pymalloc.

To improve the user experience, add a badge with the allocation and a
tooltip explaining the situation.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: Pradyun Gedam <pgedam@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>