Collect memory snapshot records in the tracker
In order to construct a new reporter that is able to plot the resident
size of the application under tracing over time, track the resident size
of the application at regular intervals when tracking as augmenting
every allocation to get timing information will have a prohibitively big
impact on the size of the result files. This new memory tracking will be
done in a background thread to not interfere with the application when
reading the relevant information from the `/proc/` pseudo-filesystem.