Update the record readers to deal with the new memory record
To allow to programmatically fetch the new memory records that we are
retrieving at regular intervals, modify the record readers to be able to
collect these new records and add APIs to retrieve them.

As we now have two main different kind of records (Allocation records
and Memory records) the reader API needs to be slightly modified to
collect both of these as the file is parsed and place them in containers
that can be retrieved later. What used to be the API to get the new
allocation record now "advances" the file parsing and can signal the
presence of either a new allocation record or a memory record. These can
be retrieved from the collections using two new APIs.