Support resetting Python stack tracker after fork
Give the thread-local Python stack tracker a method to be called from
the child process after `fork()` completes, to reset it to a sane state.

In particular, there are no longer any pending pops (because no push
records have been written to the new output file), and likewise none of
the frames on the stack have been emitted to the file yet.