Add a new `stats` reporter
Add a new reporter that allows the user to obtain some basic statistics
from a result file, including a histogram of allocation sizes, the
distribution of allocators, the top allocating locations by size and by
number of allocations and some more.

This will allow users to get basic cumulative information about how the
traced application allocates memory and where the hotspots for memory
pressure may be.

Authored-by: Subbu Ng <snochurganes@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>