tests: Make mmap tests less flaky
These tests incorrectly assume that if they use `mmap` to map and unmap
some pages before tracking begins, that those pages will still be free
later on. This is obviously an invalid assumption, since nothing has
ever guaranteed that something in the tracker or something in the
interpreter wouldn't map something else to those pages once they've been
unmapped.

Make this less flaky by moving the unmapping closer to where we hope to
re-map those pages. This still isn't bulletproof, though. It would be
better to make these unit tests that drive the `RecordReader` on an
artificially constructed capture file, rather than integration tests
that rely on being able to map new pages at the same address as old ones
in order to construct the scenario in the capture file that these tests
are trying to exercise.