Remove buffering from `RecordWriter`
This will be pushed down to be the sink's responsibility instead. The
`FileSink` will use memory mapped IO instead of an in-memory buffer that
needs to be flushed, in order to give better behavior if the process
dies to a signal or `_exit` call. The `SocketSink` will perform its own
buffering.