Don't save two copies of RecordWriter parameters
I wrote some code that tried to use `d_command_line` and it crashed,
because it is a reference that refers to a temporary that had gone out
of scope by the time I needed it.

Presumably this was a mistake and we meant to just save a copy of that
string instead of a reference, but considering that we're already saving
a copy of it inside the `d_header` struct, there's no reason to make
another copy of it as a different member variable of `RecordWriter`.

Instead, for the two member variables that are duplicated as fields in
`d_header`, just refer to `d_header` for their values whenever we need
to know them.