Guard against a null tracker singleton in hooks
Make the hooks resilient against the case where the tracker singleton is
a null pointer. It sort of worked even if they were called on a null
singleton, because most of these methods checked a static `isActive`
variable before accessing any member variables, but this is still
undefined behavior and we should still avoid it.