Simplify TLS organization for Python stack tracker
Rather than having a class with many static methods that access
thread-local state variables, make a regular class with regular methods
and regular instance variables that we can declare as thread-local.

Use a `static_assert` to ensure that it has a trivial destructor.

Access the vector through a pointer that is set when it's constructed
and unset when it's destroyed. Encapsulate the class that does the
setting within the single function that is allowed to create the vector
when it doesn't already exist.

In addition to this version providing a more idiomatic interface to
users, it's also safer than the current approach. The `static_assert`
allows us to enforce our crucial invariant that the only piece without
a trivial destructor is the vector, and scoping the struct that manages
the vector to the one function that is ever supposed to create the
vector prevents anyone from accidentally creating it by accessing it
somewhere where they shouldn't have.

Authored-by: Matt Wozniski <mwozniski@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>