Fix floating point handling in histogram binning
This code incorrectly made the assumption that, given

    step = (high - low) / bins

That you could add `step` to `low` `bins` times and wind up with a final
value that is `<= high`, but because of rounding this isn't the case.
For example:

    >>> low = 3.4657359027997265
    >>> high = 13.865115628102863
    >>> bins = 10
    >>> step = (high - low) / bins
    >>> x = low
    >>> for i in range(bins):
    ...     x += step
    ...
    >>> x
    13.865115628102867
    >>> high
    13.865115628102863
    >>> x <= high
    False

Correct this to always loop exactly `bins` times, even if that pushes
the final value over `high`.