Print example reporter command only in parent
If a process exits normally after `fork()` is called, rather than
calling `_exit`, then we will exit our tracker block twice, in both the
parent process and the child. In order to avoid confusion, only print
our message that gives an example of how to run a reporter on the output
file once, when the parent process exits and is done writing to that
output file.