Resolve Python frame line numbers in the tracker
Because we discover Python frames as they're entered, we don't know what
line number we'll wind up at within a given Python frame until the next
Python frame is entered, when we can look backwards and see what line
number the call into the new frame occurred on in the previous frame.

Up until now, we've been dumping that information to the record reader
and forcing it to reconstruct the state as it reads the records, but we
can handle this just as easily, if not more easily, in the tracker.

In fact, this makes no significant difference to the tracker's
performance, but speeds the record reader up by about 10%.