tests: Correct a test that claimed to use a thread
The name of this test is `test_hybrid_stack_in_a_thread`, but it does
not spawn a thread, and as far as I can tell it never did.

Add a new function to our C extension module that calls into native
allocations from a background thread that never acquires the GIL and
exercise it with this test, since that's an interesting test case.