Capture the frame where the tracker was entered
Previously the stacks we captured only included functions we called from
the root, without including the root frame itself. This meant that we
weren't able to include the top level line under which an allocation
occurred in the stacks produced by our reporters. It also led to a weird
effect, in that two allocation records could have equal stack traces but
different stack ids: the ID differed depending on the line number in the
root, but the stack trace itself did not.