Allow /dev/null to be used as an output file
The `FileSink` can't cope with being given `/dev/null` as an output file
anymore, because it wants to `seek` around the memory mapped file, and
`/dev/null` doesn't allow seeking.

Since it's useful to be able to suppress output when profiling
(especially on WSL where disk IO performance is terrible), make the
extension module recognize if `/dev/null` (or a symlink to it) is given
as the output file, and use a sink that silently drops everything
written to it in that case.