Utilize a reference to a TLS variable
By grabbing a reference to the TLS variable and reusing it, we can
guarantee that the TLS variable only needs to be looked up once, rather
than several times in a row. The compiler might not be able to perform
this optimization, at least when `NO_PENSIEVE_FAST_TLS` is used.