Restructure `FrameTree` to benefit `RecordReader`
By exposing a new method from `FrameTree` for resolving a child stack
index given a parent stack index and a child frame id, we can
drastically speed up the record reader. Previously, each time an
allocation record was received, the record reader would need to start
a search from the root of the frame tree down to the frame where the
allocation occurred. Now, the reader can resolve the stack ids as frames
are pushed and popped, and it only ever needs to descend one level from
the last stack ID it knew, instead of needing to start the search over
from the tree root.

Some back of the napkin profiling suggests that this makes the record
reader about 50% faster, as this tree walking was where the record
reader would spend the majority of its time.