Add a lock to protect "d_native_frames"
If the background thread is active and native frames are being
retrieved, the reader can race while adding new frames to the native
frame map while the background thread is reading the map to get the
stack trace of an element.

We need a lock in a similar way the frame map needs it.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: Matt Wozniski <mwozniski@bloomberg.net>
Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>