tests: Fix a flaky test
This test was treating every `free()` for the same address as though it
was freeing a block of memory we care about, instead of only the first
`free()` for that address after it was allocated.

Fix this by only paying attention to the first free found for an address
after it was allocated by an allocator we're interested in.

Authored-by: Matt Wozniski <mwozniski@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>