Don't rely on architecture information to get the type of the PLT table
The PLT/Jump table can have different entry types depending on the phase
of the moon, the position of the planets, the current weather and other
unpredictable stuff. Normally x86_64 uses RELA entries, and x86 uses REL
entries. But sometimes it doesn't happen, so we need to check the
DT_PLTREL tag to see which one we should use at runtime.

This method is more reliable than assuming that 64 bits means RELA and
32 bits mean REL.