docs: Improve documentation structure
In addition to some rephrasing and polishing of the wording of the
documentation, overhaul the structure. Remove the dedicated pages
showing how to run each of the CLI commands, instead merging them into
a section on the most closely related page. Create a new page for the
`run` subcommand that describes all its possible modes, but link to
a separate page for detailed descriptions about "live" mode. Move the
"live mode" page into the reporters section, since it's essentially
a reporter.