Add a `memray` entry point
Currently we only have `memray3.X` as an entry point but having also
`memray` as an entry point will make it easier for people with only one
Python version installed and it would make the usage examples from -h
line up with how we actually expect it to be used.

This commit also updates the docs and the README file to simplify the
examples with the new entry point.