Fetch line numbers for Python frames lazily
Up until now we've been greedily capturing the line number of each
Python frame as we first encounter it, but this is unnecessarily
inefficient. Since we've already been assuming that it's safe to call
`PyFrame_GetLineNumber` without the GIL held (and we'd already need to
change this code should that assumption become false in a future CPython
version), we can delay calculating the Python line number until we're
about to emit the frame. Since many frames are entered without ever
being emitted (because no allocation is performed under them), this
drastically reduces the number of calls to `PyFrame_GetLineNumber`,
which is a deceptively slow function.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>