Add `--wheel` to the `make dist` target
When we build the distribution without "--wheel", `build_ext` gets
called during "python -m build" from the temporary build directory,
while with "--wheel" the files built during "make build" are copied into
the created wheel. This allows the build to be more reproducible as in
this way the DWARF included in the extension to not change due to the
temporary directory path being part of it.