Rename `SocketDestination` attributes
To make it clearer what the effect of using a `SocketDestination` is,
use `server_port` rather than `port` and `address` rather than `host` as
the names of its attributes. Also, correct the type hint for `address`:
it mistakenly indicated that it could be `None`, but it can't (though it
can be omitted, because it has a default value).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>