Stop storing records in RecordReader
Stop making the RecordReader hold a collection of previously seen
allocation records and memory records. Instead, have it store only the
most recently seen of each.

For now, we'll just save the records in the FileReader, but this
refactoring clears the way for us to do something more memory efficient.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>