Use varint encoding for native frame info
Protobuf uses a format called "varint" for encoding integers, designed
so that smaller numbers take fewer bytes to represent. If a byte has the
most significant bit set, it indicates that there are more bytes to
follow, and if it doesn't, it indicates that you've reached the final
byte. The other 7 bits of each byte represent the value directly, with
the least significant byte first and the most significant byte last.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>