Ensure that all the addresses in the ELF tables are relocated
Some versions of libc (including some old `glibc` and `muslc`) do not
relocate the address of the ELF tables in the dynamic section. This
leads to invalid pointers when we tried to use them. To avoid this
problem we need to relocate them manually if we detect that the
addresses are incorrect. We are already doing this for the string table,
but other tables (including the GNU hash table) also need this
treatment.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>