Only use 'initial-exec' TLS model if `glibc` is detected
Using 'initial-exec' as a TLS model is not technically possible on
shared objects, but we know that `glibc` has some extra static TLS space
precisely for this purpose. In other `libc` implementations like `muslc`
, using this model is not possible and leads to relocation errors. For
this reason, fallback to 'local-dynamic' if the fast TLS model is
requested but `glibc` is not detected.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>