Add recursion guards to the `valloc`
Technically, is possible that some allocators call other allocators
underneath them. In this case we will incorrectly track both to return
the same pointer. This is generally not a problem because the many
`libc` implementatios call private symbols underneath that we are not
overriding. But this is unfortunately not always true as `muslc` uses
`memalign` underneath `valloc` without private symbols. To avoid
tracking the second call, we need to use a recursion guard.

We only do this (for now) in this call to avoid the extra overhead of
the recursion guard in the rest of the allocators (that are more
commonly used by far). We will revisit this if we found that this
happens in more places or in more versions.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>