Add an Alpine Linux step to the CI
To ensure that our package builds in Alpine Linux (with `muslc`), add a
step to the CI that tests the package in Alpine Linux. As there aren't
any version-specific concerns here (the only thing different is the
distribution and the libc implementation), just test this with the
latest Python version available.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>