Fix popping more than 16 frames with `popFrames`
Fix a bug introduced in:

    d168b0a Store count of frames to pop in FRAME_POP flags

which fortunately never made it into a released version. That commit
changed our tracker to pack the number of frames being popped into
a bit field, and handle the case where the number of frames to pop was
too large for the bit field by looping and emitting multiple records.

But, the check to see whether the number of frames to pop was too large
for the bit field was incorrect - the bit field is only 4 bits wide, and
so can only hold a maximum value of 15, not 63.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>