tests: Disable coverage on multiprocessing tests
The `pytest-cov` plugin doesn't support `multiprocessing.Pool.__exit__`.
They suggest using `Pool.close()` and `Pool.join()` instead, but that's
not an option here - what we're testing is our own behavior when the
pool is terminated, and so using `Pool.close()` and `Pool.join()` would
defeat at least one of the goals of these tests.

We may be able to solve this some other way, but for now, just disable
coverage analysis for these tests to fix our CI flakiness.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>