Use LTO and hide our symbols
This seems to achieve some minor tracking speedups on the order of 2%.

For Python 3.8 and earlier it's not as easy to hide our symbols, since
Python doesn't set the visibility for the module initialization function
to "default" automatically and Cython doesn't give us a way to control
this. For now, just skip hiding our symbols on Python 3.8 and earlier.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>