Allow tracing the Python allocators
The Python allocators don't generally generate allocation events to the
system allocator. This is how Python normally runs and is directly
responsible of keeping it fast and this translates to profiling as well.
Sometimes, users may still be interested on the requests that the
interpreter does to the Python allocators and although this generates a
lot more data and makes profiling slower, it may help in some
situations. This acts also as an alternative way to run with
`PYTHONMALLOC=malloc` but without collating the data with the system
allocator.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>