Do not aggregate native allocations located in different native lines
When we are calculating the high watermark allocation records, we
aggregate together allocations that happen in the same location.
Unfortunately, currently we define the "location" only taking into
account the Pythons tack. This means that two allocations that happen
under the same Python stack but with different native stacks underneath
will be incorrectly aggregated together leading to incorrect final
records and incorrect reports.

We must take into account both the C and the Python stack in the data
structures that are used in the snapshot to be able to produce correct
aggregated allocations.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>