Uodate the test suite to the latest version of 'rich'
Rich introduced some formatting changes in version 11.0.0 and we pinned
the version so the test suite didn't break with the new formatting.
There is really no reason other than this why we couldn't run with a
newer version and we are missing a bunch of general improvements in the
library.

This commit updates the test suite to cope with the new formatting in
the latest versions of `rich`.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>