docs: Add changelog to documentation
In order to allow Sphinx to process the changelog, change the underline
characters used by Towncrier to match the format used by the rest of our
documentation (otherwise, Sphinx tries to make one table of contents
entry for each version).  Add a symlink to `NEWS.rst` as
`docs/changelog.rst`, and indent the body of the ReST comment at the top
of the file.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>