Improve the docs for `--no-compress`
`sphinx-argparse` wanted to document `--no-compress` as having a default
value of `True`, which isn't correct. I think it's confused by the fact
that `--no-compress` has a `dest="compress_on_exit"`, which another
argument sets up with `default=True`.

Regardless, the reason we gave these things different names was so that
we could support multiple different compression options in the future.
Pre-factor this into a mutually exclusive group of compression related
options, and have `--no-compress` toggle off the one that defaults to
`True`. We would have needed to do something like this in the future if
we ever added another compression option, and this gets the Sphinx docs
looking the way we want now.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>