Add a progress bar indicator for record processing
As record processing is an operation that can take an arbitrary amount
of time depending on the size of the file, some users can be frustrated
because they don't know how much time the processing is going to take.

To help with this, add a progress bar indicator that periodically
updates while we are processing records in the `FileReader`. The class is
updated strategically to avoid impacting the performance of the actual
processing.

Co-authored-by: Matt Wozniski <godlygeek@gmail.com>
Signed-off-by: Pablo Galindo <pablogsal@gmail.com>