Allow Sinks to learn about tracker stats updates
By giving the Sink classes a new method for both writing the
`TrackerStats` and returning a pointer to a new object, they can learn
about updates to those stats that are made later. In the case of the
`FileSink`, we leverage this to allow the stats to update live, as the
process is running, to ensure that even if the process is killed by
a signal we still have accurate stats in the header.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>