Recognize and handle truncated capture files
Make the `RecordWriter` finish a capture file by writing a non-zero
trailer byte that the `RecordReader` will ignore. When the `FileSource`
reads a (non-compressed) capture file, have it ignore any trailing zero
bytes without returning them to the caller. If the caller attempts to
read into them, treat it as though EOF has been reached instead.

When a trailer byte is encountered by the reader, treat it as EOF,
rather than continuing to request bytes from the reader.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>