Fix progress bar for `--leaks` mode
For the progress bar to work, it needs to know how many records total
will be processed, but `--leaks` mode has just been passing the largest
possible `size_t` in. To make this work as expected, expose the count of
allocations from the `ProgressIndicator`, so that we can collect it on
our first pass and update our stats based on it for use in our second
pass.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>