Handle Python 3.11's non-entry frames
In Python 3.11, a single call to `_PyEval_EvalFrameDefault` can evaluate
multiple Python frames. We need to account for this when constructing
the hybrid Python/native stack, by finding all of the Python frames
associated with a single call to `_PyEval_EvalFrameDefault` and inlining
them all into the hybrid call stack at that point.

This requires our tracker to capture whether or not each Python frame is
being evaluated by the same `_PyEval_EvalFrameDefault` call as its
parent frame, and for our reader to leverage this information in its
hybrid stack reconstruction.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>