Always leverage fast unwinding to obtain the native stack
Seems that libunwind just chokes if the function that calls unw_get_reg
is not inlined in aarch64. Additionally, there are some problems with
inlined functions that call libunwind's APIs in powerpc and other
architectures.

To avoid all this trouble, leverage just the fast unwinding APIs, which
will make the program counters more consistent and will eliminate the
inline shenanigans.

Additionally increase the default size for the native stack vector to
reduce the need to call multiple times into the unwinding API.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>